# src/main/java/com/musalatest/dronetest/model/Checks.java
package com.musalatest.dronetest.model;

import lombok.Data;

import javax.persistence.*;
import java.time.LocalDateTime;


@Entity
@Data
public class Checks {
    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private Integer id;

    @Column
    private String description;

    @Column
    private LocalDateTime date;
}