# src/main/resources/static/api.yaml
openapi: "3.0.0"
info:
  title: Drone API
  description: 'The Drone REST API . Please see http://127.0.0.1:8080/docs/api for more details.'
  version: v1
  contact:
    name: Demo API Support
    url: 'http://github.com/yoidel86/demoApplication'
    email: yoidel86@gmail.com
servers:
  - url: 'http://127.0.0.1:8080/v1'
    description: localhost API server
paths:
  /drones:
    get:
      summary: List all drones
      operationId: listDrones
      parameters:
        - name: state #checking available drones for loading;
          in: query
          schema:
            type: string
            enum: ["IDLE", "LOADING", "LOADED", "DELIVERING", "DELIVERED", "RETURNING"]
            description: status
      tags:
        - drones
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drones'
    post: #- registering a drone;
      summary: Create a drone
      operationId: createDrone
      requestBody:
        description: 'Drone properties'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Drone'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drone'
  /drones/{droneId}:
    parameters:
      - name: droneId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          description: droneId
    get: #- checking loaded medication items for a given drone;
      summary: Info for a specific drone
      operationId: showDroneById
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drone'
    put: #- registering a drone;
      summary: Update a drone
      operationId: updateDrone

      requestBody:
        description: 'Drone properties'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Drone'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drone'
  /drones/{droneId}/medications:
    get: #- get loaded medications;
      summary: checking loaded medication items for a given drone;
      operationId: getDroneLoadedMedicationsById
      parameters:
        - name: droneId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            description: droneId
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medications'
        '404':
          description: A drone with the specified ID was not found.
  /drones/{droneId}/battery:
    get: #- check drone battery level for a given drone;
      summary: check drone battery level for a given drone;
      operationId: getDroneBateryLevelById
      parameters:
        - name: droneId
          in: path
          required: true
          style: simple
          schema:
            type: integer
            description: droneId
      responses:
        '200':
          description: A User object
          content:
            application/json:
              schema:
                type: object
                properties:
                  battery:
                    type: integer
                    description: The battery level.
        '404':
          description: A drone with the specified ID was not found.
  /drones/{droneId}/load/{loadId}:
    post: #- check drone battery level for a given drone;
      summary: Try to asign a Drone To a Load ;
      operationId: setDroneToLoad
      parameters:
        - name: droneId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            description: droneId
        - name: loadId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            description: loadId of a pre-registered list of medications to be delivered

      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DroneLoadedResponse'
        '404':
          description: A Drone or a Load with the specified ID was not found.
  /medications:
    get:
      summary: List all medications
      operationId: listMedications
      tags:
        - medications
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medications'
    post:
      summary: Create a Medication
      operationId: createMedication
      requestBody:
        description: 'Medication properties'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Medication'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medication'
  /medications/{medicationId}:
    parameters:
      - name: medicationId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          description: medicationId
    get:
      summary: Info for a specific Medication
      operationId: showMedicationById

      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medication'
    put:
      summary: Update a medication
      operationId: updateMedications

      requestBody:
        description: 'Medication properties'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Medication'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medication'
  /loads:
    get:
      summary: List all Loads
      operationId: listLoads
      tags:
        - loads
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loads'
    post: #loading a drone with medication items;
      summary: Create a Load
      operationId: createLoad
      requestBody:
        description: 'Load properties'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Load'
            examples:
              loadsample:
                $ref: '#/components/examples/loadsample'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Load'

  /checks:
    get:
      summary: List all Checks
      operationId: listChecks
      parameters:
        - in: query
          name: pageable
          required: false
          schema:
            $ref: '#/components/schemas/Pageable'
      tags:
        - loads
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChecksPage'

components:
  schemas:
    Drone:
      type: object
      required:
        - serialNumber
      properties:
        id:
          type: integer
        serialNumber:
          type: string
          maxLength: 100
        model:
          type: string
          enum: ["LIGHTWEIGHT", "MIDDLEWEIGHT", "CRUISERWEIGHT", "HEAVYWEIGHT"]
        weightLimit:
          type: integer
          maximum: 500
        batteryCapacity:
          type: integer
          minimum: 0
          maximum: 100
        state:
          type: string
          enum: ["IDLE", "LOADING", "LOADED", "DELIVERING", "DELIVERED", "RETURNING"]
      example:
        "serialNumber": "string"
        "model": "LIGHTWEIGHT"
        "weightLimit": 500
        "batteryCapacity": 100
        "state": "IDLE"
    Drones:
      type: array
      items:
        $ref: '#/components/schemas/Drone'
    Medication:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          pattern: '^[A-Z_0-9\-a-z]+$'
        code:
          type: string
          pattern: '^[A-Z_0-9\-]+$'
        weight:
          type: integer
          maximum: 500
        image:
          type: string
      example:
        name: Medication-2
        code: CODE-4
        weight: 300
        image: cusom.png
    Medications:
      type: array
      items:
        $ref: '#/components/schemas/Medication'
    Load:
      type: object
      required:
        - medications_id
      properties:
        id:
          type: integer
        drone_id:
          type: integer

        medications_id:
          type: array
          items:
            type: integer
    Loads:
      type: array
      items:
        $ref: '#/components/schemas/Load'
    DroneLoadedResponse:
      type: object
      properties:
        drone:
          type: object
          $ref: '#/components/schemas/Drone'
        load:
          type: object
          $ref: '#/components/schemas/Load'
    Pageable:
      description: minimal Pageable query parameters
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
          minimum: 2

    ChecksPage:
      type: object
      properties:
        size:
          type: integer
        page:
          type: integer
        pages:
          type: integer
        elements:
          type: integer
        content:
          type: array
          items:
            $ref: '#/components/schemas/Checks'

    Checks:
      description: check value generated by checkBattery task
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        date:
          type: string



    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  examples:
    loadsample:
      value:
        medications_id:
        - 1
        - 2